Revert "fullscreen/realtime interface for samples (#865)"
This reverts commit 047de727deb6f94fe066c9aa9fa77244afd0a610.
