+++ pyproject.toml
@@ -130,7 +130,6 @@ dev = [
     "pytest-dotenv",
     "pytest-xdist",
     "ruff==0.7.4", # match version specified in .pre-commit-config.yaml
-    "textual-dev>=0.86.2",
     "types-PyYAML",
     "types-aiofiles",
     "types-beautifulsoup4",
