+++ src/inspect_ai/_eval/task/run.py
@@ -9,12 +9,12 @@ from typing import Callable, Literal, cast
 
 from typing_extensions import Unpack
 
-from inspect_ai._display import (
+from inspect_ai._display import display
+from inspect_ai._display._display import (
     TaskCancelled,
     TaskError,
     TaskProfile,
     TaskSuccess,
-    display,
 )
 from inspect_ai._util.constants import (
     DEFAULT_EPOCHS,
@@ -42,7 +42,6 @@ from inspect_ai.log import (
 from inspect_ai.log._condense import condense_sample
 from inspect_ai.log._file import eval_log_json
 from inspect_ai.log._log import EvalSampleLimit, EvalSampleReductions, eval_error
-from inspect_ai.log._samples import ActiveSample, active_sample
 from inspect_ai.log._transcript import (
     ErrorEvent,
     SampleInitEvent,
@@ -259,10 +258,6 @@ async def task_run(options: TaskRunOptions) -> EvalLog:
                             log_images=log_images,
                             sample_source=sample_source,
                             sample_error=sample_error_handler,
-                            fails_on_error=(
-                                config.fail_on_error is None
-                                or config.fail_on_error is True
-                            ),
                             time_limit=config.time_limit,
                             semaphore=sample_semaphore,
                         )
@@ -367,7 +362,6 @@ async def task_run_sample(
     log_images: bool,
     sample_source: EvalSampleSource | None,
     sample_error: Callable[[BaseException], EvalError],
-    fails_on_error: bool,
     time_limit: int | None,
     semaphore: asyncio.Semaphore | None,
 ) -> dict[str, SampleScore] | None:
@@ -405,7 +399,7 @@ async def task_run_sample(
     # initialise subtask and scoring context
     init_sample_model_usage()
     set_sample_state(state)
-    sample_transcript = init_subtask(SAMPLE_SUBTASK, state.store)
+    init_subtask(SAMPLE_SUBTASK, state.store)
     if scorers:
         init_scoring_context(scorers, Target(sample.target))
 
@@ -421,27 +415,8 @@ async def task_run_sample(
         timeout(time_limit) if time_limit is not None else contextlib.nullcontext()
     )
 
-    # helper to handle exceptions (will throw if we've exceeded the limit)
-    def handle_error(ex: BaseException) -> EvalError:
-        err = sample_error(ex)
-        transcript()._event(ErrorEvent(error=err))
-        return err
-
     # solver loop
-    async with (
-        semaphore_cm,
-        sandboxenv_cm,
-        active_sample(
-            ActiveSample(
-                task_name,
-                str(state.model),
-                sample,
-                state.epoch,
-                fails_on_error,
-                sample_transcript,
-            )
-        ) as active,
-    ):
+    async with semaphore_cm, sandboxenv_cm:
         error: EvalError | None = None
         try:
             async with timeout_cm:
@@ -464,38 +439,24 @@ async def task_run_sample(
             transcript()._event(
                 SampleLimitEvent(
                     type="time",
-                    message=f"Sample completed: exceeded time limit ({time_limit:,} seconds)",
                     limit=time_limit,
+                    message=f"Sample completed: exceeded time limit ({time_limit:,} seconds)",
                 )
             )
 
             # capture most recent state for scoring
             state = sample_state() or state
 
-        except asyncio.CancelledError as ex:
-            if active.interrupt_action:
-                # record eve t
-                transcript()._event(
-                    SampleLimitEvent(
-                        type="operator",
-                        message="Sample completed: interrupted by operator",
-                    )
-                )
-
-                # handle the action
-                match active.interrupt_action:
-                    case "score":
-                        # continue to scoring (capture the most recent state)
-                        state = sample_state() or state
-                    case "error":
-                        # default error handling
-                        error = handle_error(ex)
-
-            else:
-                raise
+        except asyncio.CancelledError:
+            # allow cancelled error to propagate
+            raise
 
         except BaseException as ex:
-            error = handle_error(ex)
+            # handle error (this will throw if we've exceeded the limit)
+            error = sample_error(ex)
+
+            # fire error event
+            transcript()._event(ErrorEvent(error=error))
 
         # set timeout for scoring. if the original timeout was never hit
         # then just create a new timeout_cm targeting the original
@@ -540,14 +501,7 @@ async def task_run_sample(
                                 results[scorer_name] = sample_score
 
         except asyncio.CancelledError:
-            if active.interrupt_action:
-                transcript()._event(
-                    SampleLimitEvent(
-                        type="operator",
-                        message="Unable to score sample due to operator interruption",
-                    )
-                )
-
+            # allow cancelled error to propagate
             raise
 
         except BaseException as ex:
@@ -556,13 +510,16 @@ async def task_run_sample(
                 transcript()._event(
                     SampleLimitEvent(
                         type="time",
-                        message=f"Unable to score sample due to exceeded time limit ({time_limit:,} seconds)",
                         limit=time_limit,
+                        message=f"Unable to score sample due to exceeded time limit ({time_limit:,} seconds)",
                     )
                 )
 
             # handle error (this will throw if we've exceeded the limit)
-            error = handle_error(ex)
+            error = sample_error(ex)
+
+            # fire error event
+            transcript()._event(ErrorEvent(error=error))
 
         progress()
 
@@ -610,10 +567,11 @@ def log_sample(
         )
 
     # construct sample for logging
+    sample_events = transcript().events
 
     # if a limit was hit, note that in the Eval Sample
     limit = None
-    for e in transcript().events:
+    for e in sample_events:
         if e.event == "sample_limit":
             limit = EvalSampleLimit(
                 type=e.type, limit=e.limit if e.limit is not None else -1
@@ -634,7 +592,7 @@ def log_sample(
         output=state.output,
         scores=cast(dict[str, Score], scores),
         store=dict(state.store.items()),
-        events=list(transcript().events),
+        events=sample_events,
         model_usage=sample_model_usage(),
         error=error,
         limit=limit,
