+++ src/inspect_ai/_util/ansi.py
@@ -1,0 +1,5 @@
+import os
+
+
+def no_ansi() -> bool:
+    return os.environ.get("INSPECT_NO_ANSI", None) is not None
