+++ src/inspect_ai/_util/constants.py
@@ -30,7 +30,6 @@ DEFAULT_LOG_LEVEL_TRANSCRIPT = "info"
 ALL_LOG_FORMATS = ["eval", "json"]
 DEFAULT_LOG_FORMAT: Literal["eval", "json"] = "eval"
 EVAL_LOG_FORMAT = "eval"
-DEFAULT_DISPLAY = "rich"
 LOG_SCHEMA_VERSION = 2
 SCORED_SUFFIX = "-scored"
 SAMPLE_SUBTASK = "sample"
