+++ src/inspect_ai/log/_transcript.py
@@ -6,7 +6,6 @@ from typing import (
     Any,
     Iterator,
     Literal,
-    Sequence,
     TypeAlias,
     Union,
 )
@@ -42,9 +41,6 @@ class BaseEvent(BaseModel):
     timestamp: datetime = Field(default_factory=datetime.now)
     """Time at which event occurred."""
 
-    pending: bool | None = Field(default=None)
-    """Is this event pending?"""
-
     @field_serializer("timestamp")
     def serialize_timestamp(self, dt: datetime) -> str:
         return dt.astimezone().isoformat()
@@ -72,11 +68,11 @@ class SampleLimitEvent(BaseEvent):
     type: Literal["message", "time", "token", "operator"]
     """Type of limit that halted processing"""
 
-    message: str
-    """A message associated with this limit"""
-
     limit: int | None = Field(default=None)
-    """The limit value (if any)"""
+    """The limit value"""
+
+    message: str | None = Field(default=None)
+    """A message associated with this limit"""
 
 
 class StoreEvent(BaseEvent):
@@ -148,33 +144,20 @@ class ToolEvent(BaseEvent):
     arguments: dict[str, JsonValue]
     """Arguments to function."""
 
-    view: ToolCallContent | None = Field(default=None)
-    """Custom view of tool call input."""
-
-    result: ToolResult = Field(default_factory=str)
+    result: ToolResult
     """Function return value."""
 
     truncated: tuple[int, int] | None = Field(default=None)
     """Bytes truncated (from,to) if truncation occurred"""
 
+    view: ToolCallContent | None = Field(default=None)
+    """Custom view of tool call output."""
+
     error: ToolCallError | None = Field(default=None)
     """Error that occurred during tool call."""
 
-    events: list["Event"] = Field(default_factory=list)
+    events: list["Event"]
     """Transcript of events for tool."""
-
-    def set_result(
-        self,
-        result: ToolResult,
-        truncated: tuple[int, int] | None,
-        error: ToolCallError | None,
-        events: list["Event"],
-    ) -> None:
-        self.result = result
-        self.truncated = truncated
-        self.error = error
-        self.events = events
-        self.pending = None
 
 
 class ApprovalEvent(BaseEvent):
@@ -323,7 +306,7 @@ class Transcript:
 
     def __init__(self, name: str = "") -> None:
         self.name = name
-        self._events: list[Event] = []
+        self.events: list[Event] = []
 
     def info(self, data: JsonValue) -> None:
         """Add an `InfoEvent` to the transcript.
@@ -351,12 +334,8 @@ class Transcript:
         # end step event
         self._event(StepEvent(action="end", name=name, type=type))
 
-    @property
-    def events(self) -> Sequence[Event]:
-        return self._events
-
     def _event(self, event: Event) -> None:
-        self._events.append(event)
+        self.events.append(event)
 
 
 def transcript() -> Transcript:
