+++ src/inspect_ai/model/_call_tools.py
@@ -140,7 +140,7 @@ async def call_tools(
                 truncated=truncated,
                 view=tool_call_view(call, tdefs),
                 error=tool_error,
-                events=list(transcript().events),
+                events=transcript().events,
             )
 
             # return message and event
@@ -152,36 +152,18 @@ async def call_tools(
             ), event
 
         # call tools
-        tool_messages: list[ChatMessageTool] = []
+        results: list[tuple[ChatMessageTool, ToolEvent]] = []
         for call in message.tool_calls:
-            # create pending tool event and add it to the transcript
-            event = ToolEvent(
-                id=call.id,
-                function=call.function,
-                arguments=call.arguments,
-                view=tool_call_view(call, tdefs),
-                pending=True,
-            )
+            task = asyncio.create_task(call_tool_task(call))
+            results.append(await task)
+
+        # trace and fire tool events for each result
+        for tool_message, event in [result for result in results]:
+            trace_tool_mesage(tool_message)
             transcript()._event(event)
 
-            # execute the tool call
-            task = asyncio.create_task(call_tool_task(call))
-            tool_message, result_event = await task
-            tool_messages.append(tool_message)
-
-            # trace if we are tracing
-            trace_tool_mesage(tool_message)
-
-            # update the event with the results
-            event.set_result(
-                result=result_event.result,
-                truncated=result_event.truncated,
-                error=result_event.error,
-                events=result_event.events,
-            )
-
         # return tool messages
-        return tool_messages
+        return [result[0] for result in results]
 
     else:
         return []
