+++ src/inspect_ai/util/_sandbox/docker/compose.py
@@ -8,7 +8,7 @@ from typing import Any, Literal, TypedDict, cast
 import yaml
 from pydantic import BaseModel
 
-from inspect_ai._util.display import display_type
+from inspect_ai._util.ansi import no_ansi
 from inspect_ai._util.error import PrerequisiteError
 from inspect_ai.util._subprocess import ExecResult, subprocess
 
@@ -250,14 +250,10 @@ async def compose_command(
     env = project.env if (project.env and forward_env) else {}
 
     # ansi (apply global override)
-    if display_type() == "plain":
+    if no_ansi():
         ansi = "never"
     if ansi:
         compose_command = compose_command + ["--ansi", ansi]
-
-    # quiet if display is none
-    if display_type() == "none":
-        compose_command = compose_command + ["--progress", "quiet"]
 
     # add project scope
     compose_command = compose_command + ["--project-name", project.name]
