+++ src/inspect_ai/util/_sandbox/docker/internal.py
@@ -1,5 +1,5 @@
+from inspect_ai._util.ansi import no_ansi
 from inspect_ai._util.constants import PKG_PATH
-from inspect_ai._util.display import display_type
 from inspect_ai._util.error import PrerequisiteError
 from inspect_ai.util._subprocess import subprocess
 
@@ -24,18 +24,16 @@ async def is_internal_image_built(image: str) -> bool:
 
 
 async def build_internal_image(image: str) -> None:
-    args = [
-        "docker",
-        "build",
-        "--tag",
-        image,
-        "--progress",
-        "plain" if display_type() == "plain" else "auto",
-    ]
-    if display_type() == "none":
-        args.append("--quiet")
     result = await subprocess(
-        args + [INTERNAL_IMAGES[image].as_posix()],
+        [
+            "docker",
+            "build",
+            "--tag",
+            image,
+            "--progress",
+            "plain" if no_ansi() else "auto",
+            INTERNAL_IMAGES[image].as_posix(),
+        ],
         capture_output=False,
     )
     if not result.success:
