+++ src/inspect_ai/util/_subtask.py
@@ -125,7 +125,7 @@ def subtask(
                     name=subtask_name,
                     input=log_input,
                     result=result,
-                    events=list(transcript().events),
+                    events=transcript().events,
                     type=type,
                 )
 
@@ -154,13 +154,11 @@ def subtask(
         return create_subtask_wrapper(name)
 
 
-def init_subtask(name: str, store: Store) -> Any:
+def init_subtask(name: str, store: Store) -> None:
     from inspect_ai.log._transcript import (
         Transcript,
         init_transcript,
     )
 
     init_subtask_store(store)
-    transcript = Transcript(name=name)
-    init_transcript(transcript)
-    return transcript
+    init_transcript(Transcript(name=name))
