+++ src/inspect_ai/util/_trace.py
@@ -1,10 +1,10 @@
 from contextvars import ContextVar
 
 from rich import print
-from rich.console import RenderableType
+from rich.console import Group, RenderableType
+from rich.markdown import Markdown
+from rich.panel import Panel
 from rich.text import Text
-
-from inspect_ai._util.transcript import transcript_panel
 
 
 def trace_enabled() -> bool:
@@ -29,9 +29,37 @@ def trace_panel(
       content (RenderableType | list[RenderableType]): One or more Rich renderables.
     """
     print(
-        transcript_panel(title, subtitle, content),
+        TracePanel(title, subtitle, content),
         Text(),
     )
+
+
+class TracePanel(Panel):
+    def __init__(
+        self,
+        title: str,
+        subtitle: str | None = None,
+        content: RenderableType | list[RenderableType] = [],
+    ) -> None:
+        # resolve content to list
+        content = content if isinstance(content, list) else [content]
+
+        # inject subtitle
+        if subtitle:
+            content.insert(0, Text())
+            content.insert(0, Text.from_markup(f"[bold]{subtitle}[/bold]"))
+
+        # use vs theme for markdown code
+        for c in content:
+            if isinstance(c, Markdown):
+                c.code_theme = "xcode"
+
+        super().__init__(
+            Group(*content),
+            title=f"[bold][blue]{title}[/blue][/bold]",
+            highlight=True,
+            expand=True,
+        )
 
 
 def init_trace(trace: bool | None) -> None:
