+++ tools/vscode/src/@types/log.d.ts
@@ -187,7 +187,6 @@ export type Answer = string | null;
 export type Explanation = string | null;
 export type Metadata5 = {} | null;
 export type Timestamp = string;
-export type Pending = boolean | null;
 export type Event = "sample_init";
 export type Input1 =
   | string
@@ -207,24 +206,20 @@ export type Files1 = {
 export type Setup1 = string | null;
 export type JsonValue = unknown;
 export type Timestamp1 = string;
-export type Pending1 = boolean | null;
 export type Event1 = "sample_limit";
 export type Type5 = "message" | "time" | "token" | "operator";
-export type Message2 = string;
 export type Limit1 = number | null;
+export type Message2 = string | null;
 export type Timestamp2 = string;
-export type Pending2 = boolean | null;
 export type Event2 = "state";
 export type Op = "remove" | "add" | "replace" | "move" | "test" | "copy";
 export type Path = string;
 export type From = string | null;
 export type Changes = JsonChange[];
 export type Timestamp3 = string;
-export type Pending3 = boolean | null;
 export type Event3 = "store";
 export type Changes1 = JsonChange[];
 export type Timestamp4 = string;
-export type Pending4 = boolean | null;
 export type Event4 = "model";
 export type Model2 = string;
 export type Input2 = (
@@ -258,18 +253,16 @@ export type ToolChoice = ("auto" | "any" | "none") | ToolFunction;
 export type Name6 = string;
 export type Cache = ("read" | "write") | null;
 export type Timestamp5 = string;
-export type Pending5 = boolean | null;
 export type Event5 = "tool";
 export type Type8 = "function";
 export type Id3 = string;
 export type Function2 = string;
+export type Result = string | number | boolean | (ContentText | ContentImage)[];
+export type Truncated = [unknown, unknown] | null;
 export type Title = string | null;
 export type Format = "text" | "markdown";
 export type Content5 = string;
-export type Result = string | number | boolean | (ContentText | ContentImage)[];
-export type Truncated = [unknown, unknown] | null;
 export type Timestamp6 = string;
-export type Pending6 = boolean | null;
 export type Event6 = "approval";
 export type Message3 = string;
 export type Approver = string;
@@ -281,19 +274,15 @@ export type Decision =
   | "terminate";
 export type Explanation1 = string | null;
 export type Timestamp7 = string;
-export type Pending7 = boolean | null;
 export type Event7 = "input";
 export type Input3 = string;
 export type InputAnsi = string;
 export type Timestamp8 = string;
-export type Pending8 = boolean | null;
 export type Event8 = "score";
 export type Target2 = string | string[] | null;
 export type Timestamp9 = string;
-export type Pending9 = boolean | null;
 export type Event9 = "error";
 export type Timestamp10 = string;
-export type Pending10 = boolean | null;
 export type Event10 = "logger";
 export type Name7 = string | null;
 export type Level =
@@ -310,16 +299,13 @@ export type Filename = string;
 export type Module = string;
 export type Lineno = number;
 export type Timestamp11 = string;
-export type Pending11 = boolean | null;
 export type Event11 = "info";
 export type Timestamp12 = string;
-export type Pending12 = boolean | null;
 export type Event12 = "step";
 export type Action = "begin" | "end";
 export type Type9 = string | null;
 export type Name8 = string;
 export type Timestamp13 = string;
-export type Pending13 = boolean | null;
 export type Event13 = "subtask";
 export type Name9 = string;
 export type Type10 = string | null;
@@ -688,7 +674,6 @@ export interface Store {}
  */
 export interface SampleInitEvent {
   timestamp: Timestamp;
-  pending: Pending;
   event: Event;
   sample: Sample;
   state: JsonValue;
@@ -708,18 +693,16 @@ export interface Sample {
  */
 export interface SampleLimitEvent {
   timestamp: Timestamp1;
-  pending: Pending1;
   event: Event1;
   type: Type5;
-  message: Message2;
   limit: Limit1;
+  message: Message2;
 }
 /**
  * Change to the current `TaskState`
  */
 export interface StateEvent {
   timestamp: Timestamp2;
-  pending: Pending2;
   event: Event2;
   changes: Changes;
 }
@@ -742,7 +725,6 @@ export interface JsonChange {
  */
 export interface StoreEvent {
   timestamp: Timestamp3;
-  pending: Pending3;
   event: Event3;
   changes: Changes1;
 }
@@ -751,7 +733,6 @@ export interface StoreEvent {
  */
 export interface ModelEvent {
   timestamp: Timestamp4;
-  pending: Pending4;
   event: Event4;
   model: Model2;
   input: Input2;
@@ -868,15 +849,14 @@ export interface Response {
  */
 export interface ToolEvent {
   timestamp: Timestamp5;
-  pending: Pending5;
   event: Event5;
   type: Type8;
   id: Id3;
   function: Function2;
   arguments: Arguments1;
-  view: ToolCallContent | null;
   result: Result;
   truncated: Truncated;
+  view: ToolCallContent | null;
   error: ToolCallError | null;
   events: Events1;
 }
@@ -896,7 +876,6 @@ export interface ToolCallContent {
  */
 export interface ApprovalEvent {
   timestamp: Timestamp6;
-  pending: Pending6;
   event: Event6;
   message: Message3;
   call: ToolCall;
@@ -921,7 +900,6 @@ export interface ToolCallView {
  */
 export interface InputEvent {
   timestamp: Timestamp7;
-  pending: Pending7;
   event: Event7;
   input: Input3;
   input_ansi: InputAnsi;
@@ -931,7 +909,6 @@ export interface InputEvent {
  */
 export interface ScoreEvent {
   timestamp: Timestamp8;
-  pending: Pending8;
   event: Event8;
   score: Score;
   target: Target2;
@@ -941,7 +918,6 @@ export interface ScoreEvent {
  */
 export interface ErrorEvent {
   timestamp: Timestamp9;
-  pending: Pending9;
   event: Event9;
   error: EvalError;
 }
@@ -950,7 +926,6 @@ export interface ErrorEvent {
  */
 export interface LoggerEvent {
   timestamp: Timestamp10;
-  pending: Pending10;
   event: Event10;
   message: LoggingMessage;
 }
@@ -968,7 +943,6 @@ export interface LoggingMessage {
  */
 export interface InfoEvent {
   timestamp: Timestamp11;
-  pending: Pending11;
   event: Event11;
   data: JsonValue;
 }
@@ -977,7 +951,6 @@ export interface InfoEvent {
  */
 export interface StepEvent {
   timestamp: Timestamp12;
-  pending: Pending12;
   event: Event12;
   action: Action;
   type: Type9;
@@ -988,7 +961,6 @@ export interface StepEvent {
  */
 export interface SubtaskEvent {
   timestamp: Timestamp13;
-  pending: Pending13;
   event: Event13;
   name: Name9;
   type: Type10;
